requirements.txt
discord.py
aiohttp
requests
beautifulsoup4
pandas
python-dotenv
pytz
